deb: modularize packaging implementation
