+++ lib/rock/packaging.rb
@@ -1,0 +1,7 @@
+require_relative 'packaging/jenkins'
+require_relative 'packaging/config'
+require_relative 'packaging/gem_dependencies'
+require_relative 'packaging/obs'
+require_relative 'packaging/packager'
+require_relative 'packaging/target_platform'
+require_relative 'packaging/debian'
